+++ R/Rules-validity.R
@@ -500,3 +500,39 @@ v_stopping_tdci_ratio <- function(object) {
   )
   v$result()
 }
+
+# CohortSize ----
+
+#' Internal Helper Functions for Validation of [`CohortSize`] Objects
+#'
+#' @description `r lifecycle::badge("stable")`
+#'
+#' These functions are only used internally to validate the format of an input
+#' [`CohortSize`] or inherited classes and therefore not exported.
+#'
+#' @name v_cohort_size
+#' @param object (`CohortSize`)\cr object to validate.
+#' @return A `character` vector with the validation failure messages,
+#'   or `TRUE` in case validation passes.
+NULL
+
+#' @describeIn v_cohort_size validates that the [`CohortSizeRange`] object
+#'   contains valid `intervals` and  `cohort_size` slots.
+v_cohort_size_range <- function(object) {
+  v <- Validate()
+  v$check(
+    test_numeric(
+      object@intervals,
+      lower = 0, finite = TRUE, any.missing = FALSE, min.len = 1, unique = TRUE, sorted = TRUE
+    ),
+    "intervals must be a numeric vector with non-negative, sorted (asc.) and unique values"
+  )
+  v$check(
+    test_integer(
+      object@cohort_size,
+      lower = 0, any.missing = FALSE, len = length(object@intervals)
+    ),
+    "cohort_size must be an integer vector of the same length as intervals, containing non-negative values only"
+  )
+  v$result()
+}